Zebraix: Mostly move typing material into base.
