+++ zebraix/src/z-graph/test-files/golden-inputs/simple_graph.json
@@ -101,6 +101,15 @@
               "void"
             ]
           ]
+        },
+        {
+          "src_node": "main group",
+          "connections": [
+            [
+              "void",
+              "void"
+            ]
+          ]
         }
       ]
     }
